chore: fixtures moved to congtest + factory test in new module
